add confirm message for delete

+++ app/helpers/runs_helper.rb
@@ -8,8 +8,9 @@ module RunsHelper
 
   def user_delete(run)
     if @user == current_user
-      link_to("Delete", run_path(run), method: :delete).html_safe
+      link_to("Delete", run_path(run), method: :delete, data: {confirm: "Are you sure you want to delete this run?"}).html_safe
     end
   end
+
 end
